Метод loadUserByUsername из UserDetailsServiceImpl перенесён в отдельный класс
