+++ src/main/java/com/ncedu/eventx/services/impl/UserDetailsServiceImpl.java
@@ -1,0 +1,43 @@
+package com.ncedu.eventx.services.impl;
+
+import com.ncedu.eventx.models.entities.UserEntity;
+import com.ncedu.eventx.repositories.UserRepository;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import java.util.Arrays;
+import java.util.Collection;
+
+@Service
+public class UserDetailsServiceImpl implements UserDetailsService {
+
+    final UserRepository userRepository;
+
+    public UserDetailsServiceImpl(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        UserEntity user = userRepository.findByUsername(username);
+
+        if (user == null) {
+            throw new UsernameNotFoundException("User not found");
+        }
+
+        return new org.springframework.security.core.userdetails.User(
+                user.getUsername(),
+                user.getPassword(),
+                true, true, true, true, getAuthorities("ROLE_USER"));
+
+    }
+
+    private Collection<? extends GrantedAuthority> getAuthorities(String role) {
+
+        return Arrays.asList(new SimpleGrantedAuthority(role));
+    }
+}
